Refine inheritance model for application commands
More levels of inheritance have been added to application commands, this is to ensure that we don't end up with structures having properties that they have no actual use for.

A prime example of the above was how `(Global|Slash)SubCommand` had dud `permissions` and `defaultPermissions` properties that they had no use for. Now they do not store those anymore.