Fix error causing typealias to be expanded incorrectly in functions
We can't do `ContextCommandExecutor<Message, E, D>`  as for some reason that causes the typealias to be expanded incorrectly in functions, while it works like it should on class properties.

For example, trying to use `MessageCommandExecutor<GuildMessageEvent, GuildMessageCommandData>`  would result in a type arguments out of bounds error, stating that `GuildMessageEvent` was not a `CommandExecutorArguments`.

The same behavior applies to `UserCommandExecutor`.